environments: auto create inventory and reminder objects in the model
Before this patch the creation of reminder and inventory inside the
environment was made in the view, which made the objects created on when
addressing the http interfaces.

The problem is that it doesn't work when using the django orm directly.
(django shell or in the tests for instance).

This patch changes removes the creation from the view, and move it on
the Environment model. Now the Reminder and Inventory models are create
as soon as one creates an Environment.